requirements.txt
# Text-to-Speech
TTS>=0.15.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.22.0

# LangChain and Google AI
langchain-google-genai>=0.0.5
langchain>=0.1.0

# Environment and utilities
python-dotenv>=0.19.0
pathlib>=1.0.1